requirements.txt
feedparser
google-api-python-client
openai
python-dotenv
requests
